requirements.txt
attrs
pyyaml
requests
python-dateutil

# optional dep for redis shelf
redis